+++ common/options.go
@@ -2,11 +2,17 @@ package common
 
 import (
 	"crypto/tls"
+	"errors"
 	"github.com/go-redis/redis/v7"
 	"github.com/spf13/viper"
+	"net"
 	"os"
+	"strconv"
 	"strings"
 )
+
+const MessageMethodRedis = "redis"
+const MessageMethodDirect = "direct"
 
 type JwtOptions struct {
 	JwtSecret string
@@ -15,6 +21,7 @@ type JwtOptions struct {
 type DSockOptions struct {
 	RedisOptions *redis.Options
 	Address      string
+	Port         int
 	QuitChannel  chan struct{}
 	Debug        bool
 	LogRequests  bool
@@ -24,6 +31,12 @@ type DSockOptions struct {
 	Jwt JwtOptions
 	/// Default channels to subscribe on join
 	DefaultChannels []string
+	/// The message method between the API to the worker
+	MessagingMethod string
+	/// The worker hostname
+	DirectHostname string
+	/// The worker port
+	DirectPort int
 }
 
 func SetupConfig() error {
@@ -40,11 +53,15 @@ func SetupConfig() error {
 	viper.SetDefault("redis_db", 0)
 	viper.SetDefault("redis_max_retries", 10)
 	viper.SetDefault("redis_tls", false)
-	viper.SetDefault("address", ":6241")
+	viper.SetDefault("port", 6241)
 	viper.SetDefault("default_channels", "")
 	viper.SetDefault("token", "")
 	viper.SetDefault("jwt_secret", "")
 	viper.SetDefault("debug", false)
+	viper.SetDefault("log_requests", false)
+	viper.SetDefault("messaging_method", "redis")
+	viper.SetDefault("direct_message_hostname", "")
+	viper.SetDefault("direct_message_port", "")
 
 	err := viper.ReadInConfig()
 
@@ -63,17 +80,27 @@ func SetupConfig() error {
 	return nil
 }
 
-func GetOptions() (*DSockOptions, error) {
+func GetOptions(worker bool) (*DSockOptions, error) {
 	err := SetupConfig()
 	if err != nil {
 		return nil, err
 	}
 
-	port := os.Getenv("PORT")
+	port := viper.GetInt("port")
 
-	address := ":" + port
+	if os.Getenv("PORT") != "" {
+		println("Port env set", os.Getenv("PORT"))
+		port, err = strconv.Atoi(os.Getenv("PORT"))
 
-	if viper.GetString("address") != "" {
+		if err != nil {
+			return nil, errors.New("invalid port: could not parse integer")
+		}
+	}
+
+	address := ":" + strconv.Itoa(port)
+
+	if viper.IsSet("address") {
+		println("DEPRECATED: address is deprecated, use port")
 		address = viper.GetString("address")
 	}
 
@@ -86,6 +113,24 @@ func GetOptions() (*DSockOptions, error) {
 
 	if viper.GetBool("redis_tls") {
 		redisOptions.TLSConfig = &tls.Config{}
+	}
+
+	directHostname := ""
+	directPort := port
+	messagingMethod := viper.GetString("messaging_method")
+
+	if messagingMethod == MessageMethodRedis {
+		// OK
+	} else if messagingMethod == MessageMethodDirect {
+		if worker {
+			directHostname = viper.GetString("direct_message_hostname")
+
+			if viper.IsSet("direct_message_port") {
+				directPort = viper.GetInt("direct_message_port")
+			}
+		}
+	} else {
+		return nil, errors.New("invalid messaging method")
 	}
 
 	return &DSockOptions{
@@ -101,5 +146,25 @@ func GetOptions() (*DSockOptions, error) {
 		DefaultChannels: UniqueString(RemoveEmpty(
 			strings.Split(viper.GetString("default_channels"), ","),
 		)),
+		MessagingMethod: messagingMethod,
+		DirectHostname:  directHostname,
+		DirectPort:      directPort,
+		Port:            port,
 	}, nil
+}
+
+func GetLocalIP() string {
+	addrs, err := net.InterfaceAddrs()
+	if err != nil {
+		return ""
+	}
+	for _, address := range addrs {
+		// check the address type and if it is not a loopback the display it
+		if ipnet, ok := address.(*net.IPNet); ok && !ipnet.IP.IsLoopback() {
+			if ipnet.IP.To4() != nil {
+				return ipnet.IP.String()
+			}
+		}
+	}
+	return ""
 }